fix(src/index.ts): use the buildMiddleware function as the default export of the library
The library should be exporting a buildMiddleware function as the default export so that people can
call app.use(expressHotwire());. This'll make it more compatible and future-proof if we decide to
add options to the middleware. I've also added a test to ensure that the library is exporting the
correct thing.

BREAKING CHANGE: We are now exporting a function that builds the middleware. This matches what the
documentation has said. It also matches what other express middleware does. You should import the
middleware using app.use(expressHotwire());